+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- The server is now aware of `systemd` on Linux systems and can be started as
+  a `systemd` service with `Type=notify`.
+
 ## [2.7.0] - 2022-12-19
 
 ### Added