+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.9.2] - 2023-02-06
+
+### Fixed
+
+- Fixed MAVFTP uploads when the target path starts with `@`.
+
 ## [2.9.0] - 2023-01-18
 
 ### Added